+++ web/host_test.go
@@ -8,6 +8,7 @@ package web
 
 import (
 	"context"
+	"errors"
 	"net/http"
 	"net/http/httptest"
 	"net/url"
@@ -80,6 +81,11 @@ func TestGetSourceIp(tester *testing.T) {
 	request.Header.Set("X-real-IP", expected)
 
 	assert.Equal(tester, expected, GetSourceIp(request))
+}
+
+func TestConvertErrorToSafeString(tester *testing.T) {
+	assert.Equal(tester, "ERROR_FOO", ConvertErrorToSafeString(errors.New("ERROR_FOO")))
+	assert.Equal(tester, GENERIC_ERROR_MESSAGE, ConvertErrorToSafeString(errors.New("ERROR2_FOO")))
 }
 
 type DummyPreprocessor struct {
